refactor(card): extract card, card-recovery and fix-card-signature pages onto features/ (TASK-21854)
the 12-state card page, the recovery step machine and the signature
repair flow move into flow hooks + view components per the features/home
pattern; route pages become thin wrappers. zero behavior change.
