+++ src/features/card-recovery/components/Row.tsx
@@ -1,0 +1,8 @@
+export function Row({ label, value }: { label: string; value: string }) {
+    return (
+        <div className="flex items-center justify-between">
+            <span className="text-body-s text-foreground-secondary">{label}</span>
+            <span className="text-body-s text-foreground-primary">{value}</span>
+        </div>
+    )
+}
