+++ src/features/card/__tests__/useCardFlow.test.ts
@@ -1,0 +1,138 @@
+/**
+ * @jest-environment jsdom
+ */
+import { act, renderHook } from '@testing-library/react'
+import { ANALYTICS_EVENTS } from '@/constants/analytics.consts'
+import { rainApi } from '@/services/rain'
+import { useCardFlow } from '../useCardFlow'
+
+const mockCapture = jest.fn()
+let mockState = 'add-card'
+
+jest.mock('next/navigation', () => ({
+    notFound: jest.fn(),
+}))
+jest.mock('@tanstack/react-query', () => ({
+    useQuery: () => ({ data: undefined, isLoading: false, error: null, refetch: jest.fn() }),
+    useQueryClient: () => ({ invalidateQueries: jest.fn() }),
+}))
+jest.mock('next-intl', () => ({
+    useTranslations: () => (key: string) => key,
+}))
+jest.mock('posthog-js', () => ({
+    __esModule: true,
+    default: { capture: (...args: unknown[]) => mockCapture(...args) },
+}))
+jest.mock('@/services/card', () => ({
+    cardApi: { getInfo: jest.fn() },
+}))
+jest.mock('@/services/rain', () => ({
+    rainApi: { applyForCard: jest.fn(), getCardApplyReadiness: jest.fn() },
+}))
+jest.mock('@/services/consent', () => ({
+    cardConsentDocuments: jest.fn(() => []),
+}))
+jest.mock('@/context/authContext', () => ({
+    useAuth: () => ({ user: null, fetchUser: jest.fn() }),
+}))
+jest.mock('@/hooks/useRainCardOverview', () => ({
+    RAIN_CARD_OVERVIEW_QUERY_KEY: 'rain-card-overview',
+    useRainCardOverview: () => ({ overview: undefined, isLoading: false, error: null }),
+}))
+jest.mock('@/components/Card/cardState.utils', () => ({
+    computeCardState: () => mockState,
+}))
+jest.mock('@/components/Card/cardApply.utils', () => ({
+    pollUntilApplyAdvances: jest.fn(),
+    pollUntilReady: jest.fn(),
+}))
+jest.mock('@/app/actions/sumsub', () => ({
+    initiateSelfHealResubmission: jest.fn(),
+}))
+jest.mock('@/hooks/wallet/useGrantSessionKey', () => ({
+    useGrantSessionKey: () => ({ serializeGrant: jest.fn() }),
+}))
+jest.mock('@/hooks/useCapabilities', () => ({
+    useCapabilities: () => ({
+        railsForProvider: () => [],
+        nextActionsForRail: () => [],
+        isLoading: false,
+    }),
+}))
+jest.mock('@/hooks/useHostedVerification', () => ({
+    useHostedVerification: () => ({ start: jest.fn(), error: null }),
+}))
+jest.mock('@/context/ModalsContext', () => ({
+    useModalsContext: () => ({ setIsSupportModalOpen: jest.fn() }),
+}))
+jest.mock('@/hooks/useSafeBack', () => ({
+    useSafeBack: () => jest.fn(),
+}))
+jest.mock('@/hooks/useSumsubReloadResume', () => ({
+    useSumsubReloadResume: jest.fn(),
+}))
+
+const mockApplyForCard = rainApi.applyForCard as jest.Mock
+
+describe('useCardFlow', () => {
+    beforeEach(() => {
+        jest.clearAllMocks()
+        mockState = 'add-card'
+        window.localStorage.clear()
+    })
+
+    it('opens the sumsub sdk on an incomplete apply response', async () => {
+        mockApplyForCard.mockResolvedValue({ status: 'incomplete', sumsubAccessToken: 'tok-1' })
+        const { result } = renderHook(() => useCardFlow())
+        await act(async () => {
+            await result.current.handleApply()
+        })
+        expect(result.current.sumsubToken).toBe('tok-1')
+        expect(mockCapture).toHaveBeenCalledWith(ANALYTICS_EVENTS.CARD_SUMSUB_OPENED)
+    })
+
+    it('routes terms-required to the terms screen', async () => {
+        mockApplyForCard.mockResolvedValue({ status: 'terms-required', isUsResident: true })
+        const { result } = renderHook(() => useCardFlow())
+        await act(async () => {
+            await result.current.handleApply()
+        })
+        expect(result.current.pendingTerms).toEqual({ isUsResident: true })
+        expect(result.current.pendingCountryConfirmation).toBeNull()
+    })
+
+    it('routes country-confirmation-required before terms', async () => {
+        mockApplyForCard.mockResolvedValue({ status: 'country-confirmation-required', candidates: ['AR', 'BR'] })
+        const { result } = renderHook(() => useCardFlow())
+        await act(async () => {
+            await result.current.handleApply()
+        })
+        expect(result.current.pendingCountryConfirmation).toEqual({ candidates: ['AR', 'BR'] })
+        expect(result.current.pendingTerms).toBeNull()
+    })
+
+    it('surfaces an apply failure as applyError', async () => {
+        const consoleSpy = jest.spyOn(console, 'error').mockImplementation(() => {})
+        mockApplyForCard.mockRejectedValue(new Error('boom'))
+        const { result } = renderHook(() => useCardFlow())
+        await act(async () => {
+            await result.current.handleApply()
+        })
+        expect(result.current.applyError).toBe('boom')
+        expect(mockCapture).toHaveBeenCalledWith(ANALYTICS_EVENTS.CARD_APPLY_FAILED, { error_message: 'boom' })
+        consoleSpy.mockRestore()
+    })
+
+    it('fires the abandonment event only when sumsub closes without completing', async () => {
+        mockApplyForCard.mockResolvedValue({ status: 'incomplete', sumsubAccessToken: 'tok-1' })
+        const { result } = renderHook(() => useCardFlow())
+        await act(async () => {
+            await result.current.handleApply()
+        })
+        act(() => {
+            result.current.handleSumsubClose()
+        })
+        expect(result.current.sumsubToken).toBeNull()
+        expect(mockCapture).toHaveBeenCalledWith(ANALYTICS_EVENTS.CARD_SUMSUB_CLOSED)
+    })
+})
